pyproject.toml
[project]
name = "lumora"
version = "0.1.0"
description = "Lumora a NFT marketplace"
authors = [
    {name = "Abdul-Afeez99",email = "abdulafeezadeyemo92@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "sqlalchemy[asyncio] (>=2.0.38,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
